Remove invisible unicode from editor values in comparison
Invisible unicode characters like `\uFEFF` cause word boundaries to be added where none actually exist, and string lengths to be miscalculated.

Remove any invisible unicode control characters from text before performing comparison to ensure that the changes render correctly.
