+++ package.json
@@ -29,7 +29,7 @@
     "@asl/components": "^9.0.1",
     "@asl/constants": "^0.8.3",
     "@asl/dictionary": "^1.1.5",
-    "@asl/projects": "^9.3.6",
+    "@asl/projects": "^9.3.8",
     "@asl/service": "^8.6.2",
     "@joefitter/docx": "^4.7.0",
     "@ukhomeoffice/frontend-toolkit": "^2.13.0",
